move structs to separate file
