+++ src/contracts/mod.rs
@@ -4,9 +4,11 @@ pub mod error;
 pub mod events;
 pub mod logic;
 pub mod storage;
+pub mod structs;
 
 pub use collections::*;
 pub use entrypoints::*;
 pub use error::*;
 pub use events::*;
 pub use storage::*;
+pub use structs::*;
