+++ src/e2e/swap_route.rs
@@ -2,8 +2,8 @@
 pub mod e2e_tests {
     use crate::invariant::Invariant;
     use crate::{
-        contracts::{entrypoints::InvariantTrait, FeeTier, PoolKey},
-        invariant::{InvariantRef, SwapHop},
+        contracts::{entrypoints::InvariantTrait, FeeTier, PoolKey, SwapHop},
+        invariant::InvariantRef,
         math::types::{
             liquidity::Liquidity, percentage::Percentage, sqrt_price::calculate_sqrt_price,
             token_amount::TokenAmount,
